src/1-dsaa/16-dac/3-sort-list/sort_list.go
package main

import "fmt"

// Time complexity: TODO: change
// Space complexity: TODO: change
func sortList(input *ListNode) *ListNode {
	return input
}

func main() {
	input := []int{1, 2, 3}
	fmt.Println(sortList(arrayToListNode(input)))
}

// Boilerplate for singly linked list

type ListNode struct {
	Next  *ListNode
	Value int
}

func arrayToListNode(input []int) *ListNode {
	var headNode *ListNode
	var previousNode *ListNode

	for _, v := range input {
		nextNode := ListNode{
			Value: v,
			Next:  nil,
		}

		if headNode == nil {
			headNode = &nextNode
		} else {
			previousNode.Next = &nextNode
		}

		previousNode = &nextNode
	}
	return headNode
}

func listNodeToArray(input *ListNode) []int {
	array := []int{}
	for e := input; e != nil; e = e.Next {
		array = append(array, e.Value)
	}
	return array
}
